move gmp_cwrapper method one level up
